requirements.txt
requests==2.31.0
python-socketio==5.11.1
websocket-client==1.7.0
Flask==3.0.2
Flask-SocketIO==5.3.6